Refactor CLI helpers into SRP submodules
